requirements.txt
fastapi[standard]
uvicorn[standard]
fastapi-cors
kafka-python==2.0.2